+++ app/controllers/staff/organisations_controller.rb
@@ -12,14 +12,12 @@ class Staff::OrganisationsController < Staff::BaseController
 
     @organisation_presenter = OrganisationPresenter.new(organisation)
 
-    fund_activities = FindFundActivities.new(organisation: organisation, user: current_user).call
     programme_activities = FindProgrammeActivities.new(organisation: organisation, user: current_user).call
     project_activities = FindProjectActivities.new(organisation: organisation, user: current_user).call
     third_party_project_activities = FindThirdPartyProjectActivities.new(organisation: organisation, user: current_user).call
 
     respond_to do |format|
       format.html do
-        @fund_activities = fund_activities.map { |activity| ActivityPresenter.new(activity) }
         @programme_activities = programme_activities.map { |activity| ActivityPresenter.new(activity) }
         @project_activities = project_activities.map { |activity| ActivityPresenter.new(activity) }
         @third_party_project_activities = third_party_project_activities.map { |activity| ActivityPresenter.new(activity) }