+++ spec/features/staff/users_can_create_a_budget_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe "Users can create a budget" do
       scenario "successfully creates a budget" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
         click_on(fund_activity.title)
 
         click_on(I18n.t("page_content.budgets.button.create"))
@@ -22,7 +22,7 @@ RSpec.describe "Users can create a budget" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
 
         PublicActivity.with_tracking do
-          visit organisation_path(user.organisation)
+          visit activities_path
           click_on(fund_activity.title)
 
           click_on(I18n.t("page_content.budgets.button.create"))
@@ -43,7 +43,7 @@ RSpec.describe "Users can create a budget" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
         programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
         click_on(programme_activity.parent.title)
         click_on I18n.t("tabs.activity.details")
         click_on(programme_activity.title)
@@ -59,7 +59,7 @@ RSpec.describe "Users can create a budget" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
         programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_on(programme_activity.parent.title)
         click_on I18n.t("tabs.activity.details")
@@ -89,8 +89,7 @@ RSpec.describe "Users can create a budget" do
           parent: fund_activity,
           extending_organisation: user.organisation)
 
-        visit organisation_path(user.organisation)
-        click_on(programme_activity.title)
+        visit organisation_activity_path(user.organisation, programme_activity)
 
         expect(page).to have_content(I18n.t("page_content.activity.budgets"))
         expect(page).not_to have_content(I18n.t("page_content.budgets.button.create"))
@@ -107,9 +106,8 @@ RSpec.describe "Users can create a budget" do
           parent: programme_activity,
           organisation: user.organisation)
 
-        visit organisation_path(user.organisation)
+        visit organisation_activity_path(user.organisation, programme_activity)
 
-        click_on(programme_activity.title)
         click_on I18n.t("tabs.activity.details")
         click_on(project_activity.title)