+++ spec/features/staff/users_can_create_a_programme_level_activity_spec.rb
@@ -1,7 +1,7 @@
 RSpec.feature "Users can create a programme activity" do
   let(:user) { create(:beis_user) }
 
-  context "when signed in" do
+  context "when signed in as a BEIS user" do
     before do
       authenticate!(user: user)
     end
@@ -9,7 +9,7 @@ RSpec.feature "Users can create a programme activity" do
     scenario "successfully creates a programme" do
       fund = create(:fund_activity, organisation: user.organisation)
 
-      visit organisation_path(user.organisation)
+      visit activities_path
       click_on(I18n.t("page_content.organisation.button.create_activity"))
 
       fill_in_activity_form(level: "programme", parent: fund)
@@ -21,7 +21,7 @@ RSpec.feature "Users can create a programme activity" do
       fund = create(:activity, level: :fund, organisation: user.organisation)
       identifier = "a-programme-has-a-funding-organisation"
 
-      visit organisation_path(user.organisation)
+      visit activities_path
       click_on fund.title
       click_on I18n.t("tabs.activity.details")
       click_on(I18n.t("page_content.organisation.button.create_activity"))
@@ -38,7 +38,7 @@ RSpec.feature "Users can create a programme activity" do
       fund = create(:activity, level: :fund, organisation: user.organisation)
       identifier = "a-fund-has-an-accountable-organisation"
 
-      visit organisation_path(user.organisation)
+      visit activities_path
       click_on fund.title
       click_on I18n.t("tabs.activity.details")
       click_on(I18n.t("page_content.organisation.button.create_activity"))
@@ -55,7 +55,7 @@ RSpec.feature "Users can create a programme activity" do
       fund = create(:activity, level: :fund, organisation: user.organisation)
 
       PublicActivity.with_tracking do
-        visit organisation_path(user.organisation)
+        visit activities_path
         click_on fund.title
         click_on I18n.t("tabs.activity.details")
         click_on(I18n.t("page_content.organisation.button.create_activity"))