+++ spec/features/staff/users_can_view_budgets_on_an_activity_page_spec.rb
@@ -12,7 +12,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         budget = create(:budget, parent_activity: fund_activity)
         budget_presenter = BudgetPresenter.new(budget)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link fund_activity.title
 
@@ -25,7 +25,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         budget_2 = create(:budget, parent_activity: fund_activity, period_start_date: 1.year.ago, period_end_date: Date.yesterday)
         budget_3 = create(:budget, parent_activity: fund_activity, period_start_date: 2.years.ago, period_end_date: 1.year.ago)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link fund_activity.title
         expect(page.find(:xpath, "//table[@class = 'govuk-table budgets']/tbody/tr[1]")[:id]).to eq(budget_1.id)
@@ -42,7 +42,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         budget = create(:budget, parent_activity: programme_activity)
         budget_presenter = BudgetPresenter.new(budget)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link fund_activity.title
         click_on I18n.t("tabs.activity.details")
@@ -62,7 +62,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         budget = create(:budget, parent_activity: project_activity)
         budget_presenter = BudgetPresenter.new(budget)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link fund_activity.title
         click_on I18n.t("tabs.activity.details")
@@ -80,7 +80,7 @@ RSpec.feature "Users can view budgets on an activity page" do
 
         budget = create(:budget, parent_activity: project_activity)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link fund_activity.title
         click_on I18n.t("tabs.activity.details")
@@ -106,9 +106,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         budget = create(:budget, parent_activity: programme_activity)
         budget_presenter = BudgetPresenter.new(budget)
 
-        visit organisation_path(user.organisation)
-
-        click_link programme_activity.title
+        visit organisation_activity_path(user.organisation, programme_activity)
 
         budget_information_is_shown_on_page(budget_presenter)
       end
@@ -118,9 +116,7 @@ RSpec.feature "Users can view budgets on an activity page" do
 
         budget = create(:budget, parent_activity: programme_activity)
 
-        visit organisation_path(user.organisation)
-
-        click_link programme_activity.title
+        visit organisation_activity_path(user.organisation, programme_activity)
 
         within "##{budget.id}" do
           expect(page).to_not have_content I18n.t("default.link.edit")
@@ -136,7 +132,7 @@ RSpec.feature "Users can view budgets on an activity page" do
         budget = create(:budget, parent_activity: project_activity)
         budget_presenter = BudgetPresenter.new(budget)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link programme_activity.title
         click_on I18n.t("tabs.activity.details")
@@ -151,7 +147,7 @@ RSpec.feature "Users can view budgets on an activity page" do
 
         budget = create(:budget, parent_activity: project_activity)
 
-        visit organisation_path(user.organisation)
+        visit activities_path
 
         click_link programme_activity.title
         click_on I18n.t("tabs.activity.details")