+++ CHANGELOG.md
@@ -205,6 +205,7 @@
 - Activity creation journey changed to ask for a level and parent
 - Add an activities page and navigation
 - Activities can be filtered to an organisation
+- Fund activities are no longer shown on the home page
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-11...HEAD
 [release-11]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-10...release-11